refactor: core.repository to use Protocol
- replace abstract class with Protocol

- rename repository.py to project.py
