+++ src/core/repository/__init__.py
@@ -1,3 +1,2 @@
 # pragma: no cover
 # noqa: D104
-from .repository import *
